recovery_project/app/api_app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
httpx>=0.25.0
psycopg2-binary>=2.9.9
pgvector>=0.2.3
python-dotenv>=1.0.0
python-multipart>=0.0.6
email-validator>=2.0.0
ujson>=5.8.0
openai>=1.0.0 
langchain
langchain_community
